+++ Gruntfile.js
@@ -17,7 +17,7 @@ module.exports = function(grunt) {
   var banner =
     "/** MIT License\n" +
     "* \n" +
-    "* Copyright (c) 2015 Ludovic CLUBER \n" +
+    "* Copyright (c) 2017 Ludovic CLUBER \n" +
     "* \n" +
     "* Permission is hereby granted, free of charge, to any person obtaining a copy\n" +
     '* of this software and associated documentation files (the "Software"), to deal\n' +