+++ src/ts/types.ts
@@ -1,2 +1,3 @@
 export type LevelName = "info" | "time" | "trace" | "warn" | "error" | "off";
 export type MessageContent = string | number | any[] | Object;
+export type ConsoleMethod = "info" | "trace" | "warn" | "error";